app/requirements.txt
flask==1.0.3
gunicorn==19.9.0
pymongo==3.8.0
python-dotenv==0.10.3